:dash: move construct_update_data_patch method into the FigureResampler class
