+++ plotly_resampler/figure_resampler/figure_resampler.py
@@ -13,7 +13,7 @@ __author__ = "Jonas Van Der Donckt, Jeroen Van Der Donckt, Emiel Deprost"
 import os
 import warnings
 from pathlib import Path
-from typing import List, Optional, Tuple
+from typing import List, Optional, Tuple, Union
 
 import dash
 import plotly.graph_objects as go
@@ -656,6 +656,47 @@ class FigureResampler(AbstractFigureAggregator, go.Figure):
                 + "\t- 'show-dash' method was not called, or \n"
                 + "\t- the dash-server wasn't started with 'show_dash'"
             )
+
+    def construct_update_data_patch(
+        self, relayout_data: dict
+    ) -> Union[dash.Patch, dash.no_update]:
+        """Construct the Patch of the to-be-updated front-end data, based on the layout
+        change.
+
+        Attention
+        ---------
+        This method is tightly coupled with Dash app callbacks. It takes the front-end
+        figure its ``relayoutData`` as input and returns the ``dash.Patch`` which needs
+        to be sent to the ``figure`` property for the corresponding ``dcc.Graph``.
+
+        Parameters
+        ----------
+        relayout_data: dict
+            A dict containing the ``relayoutData`` (i.e., the changed layout data) of
+            the corresponding front-end graph.
+
+        Returns
+        -------
+        dash.Patch:
+            The Patch object containing the figure updates which needs to be sent to
+            the front-end.
+
+        """
+        update_data = self._construct_update_data(relayout_data)
+        if not isinstance(update_data, list) or len(update_data) <= 1:
+            return dash.no_update
+
+        patched_figure = dash.Patch()  # create patch
+        for trace in update_data[1:]:  # skip first item as it contains the relayout
+            trace_index = trace.pop("index")  # the index of the corresponding trace
+            # All the other items are the trace properties which needs to be updated
+            for k, v in trace.items():
+                # NOTE: we need to use the `patched_figure` as a dict, and not
+                # `patched_figure.data` as the latter will replace **all** the
+                # data for the corresponding trace, and we just want to update the
+                # specific trace its properties.
+                patched_figure["data"][trace_index][k] = v
+        return patched_figure
 
     def register_update_graph_callback(
         self,
